backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
supabase==2.3.4
gotrue==2.4.2
httpx==0.24.1
httpcore==0.17.3
pydantic==2.5.3
email-validator==2.1.0
google-generativeai